refactor: move vmm logic into wrapper struct
