+++ vmm/Cargo.toml
@@ -25,3 +25,4 @@ elf = "0.7.4"
 env_logger = "0.11.6"
 log = "0.4.25"
 rustc-demangle = "0.1.24"
+anyhow = "1.0.97"
