+++ vmm/src/lib.rs
@@ -2,5 +2,8 @@
 #![feature(ptr_metadata)]
 #![feature(maybe_uninit_write_slice)]
 #![feature(box_into_inner)]
+#![feature(str_from_raw_parts)]
+#![feature(exitcode_exit_method)]
 
 pub mod client;
+pub mod runtime;
